DAQClient/pyproject.toml

[tool.poetry]
name = "daqclient"
version = "0.1.0"
description = "A program that subscribes to new data events, reads them, and passes them on to image processors."
authors = ["Reinier van Mourik <reinier.vanmourik@tausystems.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pytango = "^9.4.2"

# image libraries
imageio = "^2.31.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
